SSCSCI-9: finish refactor of fetch secret code
